update by product id

+++ inventory.php
@@ -1,5 +1,18 @@
 <?php
     
+    if('UPDATE_BY_PRODUCTID' == $action){
+        $productid = $_POST['productid'];
+        $quantity = $_POST['quantity'];
+        $sql = "UPDATE $table SET quantity = '$quantity' WHERE productid = '$productid'";
+        if ($conn->query($sql) === TRUE) { 
+            echo "success";
+        } else {
+            echo "failed";
+        }
+        $conn->close();
+        return;
+    }
+
     if('DELETE_EID' == $action){
         $eid = $_POST['eid'];
         $sql = "DELETE FROM $table WHERE eid = '$eid'";
